+++ tests/primitives.lua
@@ -11,9 +11,9 @@ testPrimitives = {}
 function testPrimitives:testCircle()
     for i=1,10,1 do
         local circle_raster = primitives.circle(i)
-        -- Check that all enclosed points are less than i from centre
-        local inside_circle = subpattern.enclosed(circle_raster)
-        lu.assertEquals(#inside_circle, 1) -- Should only be one enclosed segment
+        -- Check that all holes are less than i from centre
+        local inside_circle = subpattern.interior_holes(circle_raster)
+        lu.assertEquals(#inside_circle, 1) -- Should only be one hole
         for pt in inside_circle[1]:cells() do
             lu.assertTrue(cell.euclidean(pt, cell.new(0,0)) < i )
         end